Brief check in Survival [temporary name of a world]
Everything seems to work fine.

A creeper blew up; filled the hole with water.

The cat in our house is more whiny than I remember haha.

Gaming on the MBPro is far from pleasant. ~7k RPM and a lot of heat under the hands. So it's a legit stereotype hahah.

> Will proceed to check YellowVillage now :)