Checked YellowVillageWorld
Oh, I've had misidentified which world this actually was. It's not the "big-fatty-wall" one, with a lot of redstone machinery and trails...it's the "yellow-wool-wall" one, the one with an underground extension of the village, with homes all built with glowstone (actually moved myself down there, and updated the world icon 🙂).

>> Should we maybe rebrand it as "Glowstone Village"? 🙃

Ah, also, everything works on the MBPro. Except again huuuuge fan noise (~7k RPM...) and quite a lot of heat around the touchbar, propagating uncomfortably up to the "wrist-rests" and the trackpad.

But I should add for fairness that CPU temperatures never crossed the 70° line; instead they always stayed between ~55° and ~65°, which seems to actually be very fair. I think we'd say that heat dissipation works good after all... 😅