Intrigue: a world may be missing [start]
Is "Survival" world actually the "big-fatty-wall-with-redstone-trails" one?
Where is that world, if not?

I have no time now to really explore, but it might be in another village somewehere. If so, we cold think of a proper name for the world starting from there...

if not: @trincallegra, where is it? Maybe one of the pruned worlds in your local .minecraft/saves folder...? We should investigate 🧐

>> For the time being I've updated the Survival world icon 🙂