Clarifying what repo-size badge gives
The actual meaning of the `repo-size` badge from shields.io (as introduced in commit 6380229) was debated, for it displays a much smaller value than the local repository size. 
After some investigation it appears that it could loosely correspond to the total size of a `bare clone` of the repository, meaning a clone of the .git files, without anything checked out in the working tree. This could make a lot of sense, since there is no real sensible criterion on which head (or generic commit) to checkout, and all of them have different size.

> So, taking this info as a candidate solution for the repo-size intrigue, we have updated the badge label, from `[git-logo] totalling` to `[git-logo] bare size`.