OCTOMAPS! [experimental]
Trying out the @githubocto workflow for drawing nice "bird-eye views" of repositories. 

I already tested in the interactive web-app (https://octo-repo-visualization.vercel.app/) how mcworlds are represented and determined that could be actually quite useful to efficiently gather visual-info about:
- how much of the overworld, nether and end dimensions has been generated 
- how much entities have been 'activated'
- how much points of interested discovered
- and so on... 
[more info at https://minecraft.fandom.com/wiki/Java_Edition_level_format]

I genuinely think the tool is generally great, and especially useful for this quite peculiar and unusual use of a git repository. Happy to play with it!